+++ ChangeLog.md
@@ -60,6 +60,9 @@ Changes in behavior:
 17. NLog is no longer enabled by default in application runtime.
     * To enable NLog for Rhetos components, in Program.CreateHostBuilder,
       add `hostBuilder.UseNLog();`, and in Startup.ConfigureServices: `.AddHostLogging()` after AddRhetosHost.
+18. Changed how navigation properties are generated in Queryable classes:
+    * Navigation properties can be set only during the initialization phase. That means that after the Queryable object is constructed the navigation properties are read only.
+    * Navigation properties behaviour can no longer be extended with custom tags *DataStructureQueryable Getter* and *DataStructureQueryable Setter*
 
 Changes in Rhetos libraries API:
 
@@ -129,6 +132,9 @@ Changes in Rhetos libraries API:
     * Instead of using `ProcessContainer.CreateScope(rhetosAppAssemblyPath)` replace it with `LinqPadRhetosHost.CreateScope(rhetosAppAssemblyPath)`
 16. Removed `IUserInfoAdmin` interface. It was used together with the `Rhetos:AppSecurity:BuiltinAdminOverride` option to give the administrator rights as it had all claims.
 17. GetInternalServerErrorMessage method moved from FrameworkException class to Rhetos.Utilities.ErrorReporting.
+18. The following methods are no longer available: EntityFrameworkContext.ClearCache(), IPersistenceCache.ClearCache() and ToNavigation(). They where Rhetos-specific helpers for saving entity with Entity Framework.
+    * If custom application code overrides repository's standard Save method, use DomHelper.WriteToDatabase to save an entity, instead of Entity Framework methods. For reference, see the new generated source code for repository Save methods.
+19. Method `DataStructureQueryableCodeGenerator.AddNavigationPropertyWithBackingField` is renamed to `DataStructureQueryableCodeGenerator.AddNavigationProperty` and the use of parameter `additionalSetterCode` is removed,
 
 ### Internal improvements